IsGranted 80%

+++ src/Controller/RegistrationController.php
@@ -13,19 +13,27 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/api/tournaments')]
 class RegistrationController extends AbstractController
 {
     #[Route('/{id}/registrations', name: 'registration_index', methods: ['GET'])]
+    #[IsGranted('ROLE_USER')]
     public function index(Tournament $tournament, RegistrationRepository $repo, SerializerInterface $serializer): JsonResponse
     {
+        $currentUser = $this->getUser();
+        if ($tournament->getOrganizer() !== $currentUser && !in_array('ROLE_ADMIN', $currentUser->getRoles())) {
+            return $this->json(['message' => 'Access denied.'], Response::HTTP_FORBIDDEN);
+        }
+        
         $registrations = $repo->findBy(['tournament' => $tournament]);
         $json = $serializer->serialize($registrations, 'json', ['groups' => 'registration:read']);
         return new JsonResponse($json, Response::HTTP_OK, [], true);
     }
 
     #[Route('/{id}/registrations', name: 'registration_create', methods: ['POST'])]
+    #[IsGranted('ROLE_USER')]
     public function create(Tournament $tournament, Request $request, EntityManagerInterface $em, SerializerInterface $serializer): JsonResponse
     {
         $data = json_decode($request->getContent(), true);
@@ -54,8 +62,15 @@ class RegistrationController extends AbstractController
     }
 
     #[Route('/{idTournament}/registrations/{idRegistration}', name: 'registration_delete', methods: ['DELETE'])]
+    #[IsGranted('ROLE_USER')]
     public function delete(Registration $idRegistration, EntityManagerInterface $em): JsonResponse
     {
+        $currentUser = $this->getUser();
+        $tournament = $idRegistration->getTournament();
+        if ($tournament->getOrganizer() !== $currentUser && !in_array('ROLE_ADMIN', $currentUser->getRoles())) {
+            return $this->json(['message' => 'Access denied.'], Response::HTTP_FORBIDDEN);
+        }
+
         $em->remove($idRegistration);
         $em->flush();
         return new JsonResponse(null, Response::HTTP_NO_CONTENT);
