+++ src/Controller/TournamentController.php
@@ -12,11 +12,13 @@ use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/api/tournaments')]
 class TournamentController extends AbstractController
 {
     #[Route('', name: 'tournament_index', methods: ['GET'])]
+    #[IsGranted('ROLE_USER')]
     public function index(TournamentRepository $repository, SerializerInterface $serializer): JsonResponse
     {
         $tournaments = $repository->findAll();
@@ -25,6 +27,7 @@ class TournamentController extends AbstractController
     }
 
     #[Route('/{id}', name: 'tournament_show', methods: ['GET'])]
+    #[IsGranted('ROLE_USER')]
     public function show(Tournament $tournament, SerializerInterface $serializer): JsonResponse
     {
         $json = $serializer->serialize($tournament, 'json', ['groups' => 'tournament:read']);
@@ -32,6 +35,7 @@ class TournamentController extends AbstractController
     }
 
     #[Route('', name: 'tournament_create', methods: ['POST'])]
+    #[IsGranted('ROLE_USER')]
     public function create(Request $request, EntityManagerInterface $em, SerializerInterface $serializer): JsonResponse
     {
         $data = json_decode($request->getContent(), true);
@@ -59,8 +63,14 @@ class TournamentController extends AbstractController
     }
 
     #[Route('/{id}', name: 'tournament_update', methods: ['PUT'])]
+    #[IsGranted('ROLE_USER')]
     public function update(Request $request, Tournament $tournament, EntityManagerInterface $em, SerializerInterface $serializer): JsonResponse
     {
+        $currentUser = $this->getUser();
+        if ($tournament->getOrganizer() !== $currentUser && !in_array('ROLE_ADMIN', $currentUser->getRoles())) {
+            return $this->json(['message' => 'Access denied.'], Response::HTTP_FORBIDDEN);
+        }
+
         $serializer->deserialize($request->getContent(), Tournament::class, 'json', [
             'object_to_populate' => $tournament,
         ]);
@@ -71,8 +81,14 @@ class TournamentController extends AbstractController
     }
 
     #[Route('/{id}', name: 'tournament_delete', methods: ['DELETE'])]
+    #[IsGranted('ROLE_USER')]
     public function delete(Tournament $tournament, EntityManagerInterface $em): JsonResponse
     {
+        $currentUser = $this->getUser();
+        if ($tournament->getOrganizer() !== $currentUser && !in_array('ROLE_ADMIN', $currentUser->getRoles())) {
+            return $this->json(['message' => 'Access denied.'], Response::HTTP_FORBIDDEN);
+        }
+        
         $em->remove($tournament);
         $em->flush();
         return new JsonResponse(null, Response::HTTP_NO_CONTENT);
